unit tests refactor
